Build a list of all candidates in determineManifestConversion
Instead of trusting ImageDestination.SupportedManifestMIMETypes() to be
perfectly correct, build a list of all possible MIME types to try.

This will be useful for supporting both schema1+schema2 and schema1-only
docker/distribution registries, which can’t be (at least in the case of
OpenShift) distinguished without actually trying to upload a manifest.

This is not used yet, the user will be added in the next commit.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>