Always use serverDefault() for docker TLS configuration
Previously we would not use it if there was any other TLS configuration.

If the user specified certificates, this really does not make any sense
(there is no overlap between serverDeafult() and certificates).

If the user disabled TLS verification, skipping serverDefault() is a bit
more sensible (e.g. we may want to allow old ciphers because it doesn’t
matter anyway), but still suspect.  If this turns out to be a problem,
the DockerInsecureTLSSkipVerify case can revert the specific settings,
making the intention explicit.

This also nicely simplifies the code.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>