storage: add support for manifest list
If `SupportedManifestMIMETypes` returns `nil` in `copy.Image` we don't
handle manifest conversions from docker manifest list. This patch adds
docker's mime types so that when a manifest list comes from a source
image, `copy.Image` converts it to a single manifest so it can go ahead
and actually copy the image instead of erroring out.

Signed-off-by: Antonio Murdaca <runcom@redhat.com>