Fix checks for storage.Errors and test for them
The storage library can return its errors wrapped by
github.com/pkg/errors.Wrap() and friends, so when checking for equality,
use errors.Cause() to dig into them.

Tests for duplicate ID detection were broken starting in
3df63b469375caa5762b54f6056e51446129779c, which made that acceptable if
the new image and the old image had the same top layer, so adjust the
tests for that error to use different layer contents when creating the
second image with the same ID as one that it had just created.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>