Support v3 signature packets in mechanism_openpgp
They are handled similarly to the more common V4 signature packets,
except that MessageDetails.SignatureV3 is set instead of
MessageDetails.Signature, and they don’t support expiration.

Bulk of the commit is adding tests which use V3 signature packets.

I have regenerated unknown-key.signature because it was generated using
a temporary key created just for that purpose, and the test is simpler
when fixtures for both packet variants to use the same key.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>