Add registries.conf file and build library to parse it
registries.conf is a system-wide configuration file initially
developed in https://github.com/projectatomic/registries.  We
eventually want to ship the configuration file with this project;
hence the addition.  However, we also needed a library like
approach that gave this and other projects the ability to
parse the file in the same manner.

Signed-off-by: baude <bbaude@redhat.com>