Make sure TMPDIR does not contain symlinks
Some tests directly or indirectly exercising the directory/explicitfilepath
subpackage expect the path returned by ioutil.TempDir to be canonical in the
directory/explicitfilepath sense (absolute, no symlinks, cleaned up).

ioutil.TempDir uses $TMPDIR by default, and on macOS, $TMPDIR is by
default set to /var/folders/…, with /var a symlink to /private/var ,
which does not match our expectations.  So, tests which want to use
ioutil.TempDir that way, can
import _ "github.com/containesr/image/internal/testing/explicitfilepath-tmpdir"
to ensure that $TMPDIR is canonical and usable as a base for testing
path canonicalization in its subdirectories.

Hide this canonicalization in a subpackage which can be just imported into
any tests which needs it.

(This also introduces the first use of the "internal" subdirectory,
for subpackages which should not be used by externals callers.)

Signed-off-by: Miloslav Trmač <mitr@redhat.com>