Remove types.Image.IsMultiImage
Instead, have copy.Image call .Manifest() and use manifest.MIMETypeIsMultiImage
directly.

This is not locally an improvement; but we want to move the IsMultiImage check
to work on types.UnparsedImage, and because types.Image is a superset of
types.UnparsedImage, for consistency we would have to implement a
completely useless sourcedImage.IsMultiImage which works with the cached data.

Instead, let's just drop the method from the types.Image interface and work
with the MIME type directly.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>