Rename copy.imageCopier to copier
The copier name is not that great, but we will want separate objects
for a per-manifest-list state (e.g. information about blobs, which
may be shared) and per-image-instace state (e.g. manifest conversions).
For lack of better ideas, copier will be the per-manifest-list one,
and imageCopier the per-image-instance one.

For now this commit only renames the old copier object, splitting
it into copier/imageCopier will happen later.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>