Use copier and imageCopier instead of parameters and variables
Make more functinos methods of copier/imageCopier, and use fields
of these objects instead of local variables in copy.Image.

This primarily makes it clear which values are associated with which
concepts and state, and simplifies rearranging the large copy.Image.

This makes the lifetime of some values a bit more opaque, but IDE
reference lookup tools are eaiser to use with only one copy of a value
to track, which is roughly a wash.

The one real downside is that tests are now creating mock copier/imageCopier
objects with only the fields necessary to test a particular function,
i.e. the objects are mostly invalid.  Perhaps that can be cleared up later
with better initialization helpers and some common "unusable ImageSource" mocks.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>