If the copy.Image source is a manifest list, copy a single instance
Finally: If the top-level image is a manifest list, select an appropriate instance,
create a new types.UnparsedImage for the instance, and copy only that instance.
The instanceDigest handling in UnparsedImage and dependencies will take care
of handling the rest.

For several transports we should instead copy all of the instances in the manifest
list; this is not implemented yet, but at least the source instance access
part is ready.  (Remaining: listing images in a manifest list,
instanceDigest handling in destinations, updating and copying the manifest list.)

Signed-off-by: Miloslav Trmač <mitr@redhat.com>