Add architecture/OS overrides to types.SystemContext
... and use them for choosing an image from a manifest list, and
verifying whether an image is acceptable in a MustMatchRuntimeOS()
destination.  Propagate the types.SystemContext through the
call stack as necessary.

This adds no users and seems not all that important, but after
we re-enable fetching manifest lists, docker_transport_test.go does a
"//busybox".NewImage(), which nowadays mean copying from a manifest list,
and then we need this override to keep tests working on non-Linux platforms
for which Docker does not publish images.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>