Drop lack of DType support on overlay to warning
Currently RHEL7.4 or later OS is required to support DType well.
We want to eventually make this an error if you select overlay on
a file system without DType, but for now in order to allow
containers/storage to get merged into CRI-O we need to relax this
constraint since the Test infrastucture does not have support.

Once we have the support we can revert this change.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>