Specify OverrideKernelCheck and Size in /etc/containers/storage.conf
Allow users to specify the maximum size of a containers in the
/etc/containers/storage.conf file.  This will enable quota checking
in the drivers that support it.

Allowe users to OverrideKernelCheck to tell drivers to ignore kernel
version checks that indicate whether the driver is supported on that
kernel.  These checks do not make sense on older RHEL kernels, since
the driver support was back ported.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>