Add a "tarball" transport for importing tarballs
Add a "tarball" transport which can be used to import tarballs of root
filesystem as images by treating each tarball as a single layer,
supplying a default OCI configuration (or one passed to the image from
which the ImageSource or Image was initialized), and generating an OCI
manifest to describe it all.

A tarball named "-" is buffered in memory using the contents of stdin.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>