overlay: when checking for support, actually try a mount
When checking if we can use the overlay driver at initialization time,
actually try a mount using temporary directories instead of assuming
that the kernel listing it as a known filesystem is enough, since it
might not be listed if we're in a container, and we can't load the
module from inside of a container.

We try first with two lowers.  If two lowers doesn't succeed, but we
succeed with just one, return a "kernel too old to provide multiple
lowers feature" error.  If one does, then return a "not supported over
<filesystem> at <homedir>" error.

Do the supports-d_type check at the same time, since that lets us avoid
having to create another directory to try reading to check for it.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>