Remove .tar extension from blob and config file names
The config file was being saved as digest.tar for the directory transport.
This is misleading as the config file is not a tar archive.
Dropped the .tar extension from all files now, including blobs.
The user can use a tool like file to determine the format of the files in the directory.

Signed-off-by: umohnani8 <umohnani@redhat.com>