overlay: always mount, even if we have no parent layer
Always mount an overlay filesystem, even when there's no parent layer,
so that we can set an SELinux context for the mountpoints.  We create an
"empty" directory to use as the lower directory for the overlay mount to
satisfy the kernel, but as it's not a real layer, we don't track it in
the lists of lower directories.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>