changes: check for whiteout when looking at lower layers
When deciding whether to mark a file as added or modified, we check for
the presence of an item with the same path in a layer from which the
current later might have inherited a different version of the file.
This can be tripped up by overlay-style whiteout entries, which have the
same name as the file which they mark as having been removed.  Also make
sure that we don't mistakenly count, as inherited, a file with the same
name that was in a lower layer, but which was removed by an intermediate
layer.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>