Support gzip for docker-archive files
The docker-archive transport pulls in images from tar file
but would fail if the tar file was compressed using gzip.
This adds the functionality of detecting whether the file is
compressed, and if it is the contents are copied to a temp
uncompressed file.
This was supported for the docker-daemon transport, so moving
that functionality to tarfile/src.go fo both docker-archive
and docker-daemon.

Signed-off-by: umohnani8 <umohnani@redhat.com>