Rewrite TestTransportValidatePolicyConfigurationScope
For some unfanthomable reason I wrote the test to accept non-canonical
inputs to ParseReference, instead of the canonicalized prefixes returned by
PolicyConfigurationIdentity and PolicyConfigurationNamespace.

So, replace the expected cases by the actually returned forms, up to
the full name:tag@digest@ID, and the relevant prefixes.

This only changes the tests; the cases which would fail are commented out
for now.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>