Fix ValidatePolicyConfigurationScope
Re-enable the failing tests, and rewrite the implementation to match expected inputs.

Notably this re-enables the @ID form with no name.

Like docker.Transport, this does no validation of the name part; using
the docker/reference parses is not correct because they are not intended to
accept e.g. hostname-only "docker.io"

Signed-off-by: Miloslav Trmač <mitr@redhat.com>