Simplify ParseStoreReference a bit more
Now that refname == completeReference.String(), use the method call
instead of the variable where convenient.  A little later we will
eliminate the variable entirely, for now it is still useful
to pass to newReference.

Also drop a comment which became misleading: the code following it is there
primarily to incorporate a digest into the value, only secondarily to format
a string.  (And even the digest incorporation will hopefully go away.)

Signed-off-by: Miloslav Trmač <mitr@redhat.com>